+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 0.2.10 (WIP)
 - FIX: Reverted hotfix from 0.2.9 and implemented validation to use iso yaml file only for editable package installs and emit warning
+- FIX: `-o` output refactored to no longer expect the result to be structured by support levels
+- TWEAK: Refactored multiple file input result interesction and union
+- TESTS: Added tests for multiple file input intersection and union results
 
 ## 0.2.9 (24.03.2021)
 - HOTFIX: Prevent error message about missing file in CLI use